dota with isaid to satellitepy
I tried to fix the error some more but could not get to a solution.
I cant imagine that this is the problem, but the error occures at file P0019 of the validation files. I looked into the corresponding mask file and some of the masks overlap and I found one thats totally inside another. Maybe it cant find that one which would lead to one more bbox than masks. But that wouldnt explain why it works when i first create all the patches and add the mask coordinates afterwards